# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get all articles or search based on filters",
                "operationId": "getArticles",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "Keyword to search for articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for filtering articles",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for filtering articles",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Category to filter articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Source to filter articles",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of articles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/articles/personalized-feed": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get the personalized news feed for the authenticated user",
                "operationId": "getPersonalizedFeed",
                "responses": {
                    "200": {
                        "description": "Personalized news feed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "articles": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Article"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Article": {
                "title": "Article",
                "description": "Article model",
                "properties": {
                    "id": {
                        "description": "The unique identifier for the article",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "The title of the article",
                        "type": "string",
                        "example": "Sample Article"
                    },
                    "description": {
                        "description": "The description of the article",
                        "type": "string",
                        "example": "This is a sample article description"
                    },
                    "content": {
                        "description": "The content of the article",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet, consectetur adipiscing elit..."
                    },
                    "author": {
                        "description": "The author of the article",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "source": {
                        "description": "The source of the article",
                        "type": "string",
                        "example": "New York Times"
                    },
                    "category": {
                        "description": "The category of the article",
                        "type": "string",
                        "example": "Technology"
                    },
                    "url": {
                        "description": "The URL of the article",
                        "type": "string",
                        "example": "https://example.com/article"
                    },
                    "url_to_image": {
                        "description": "The URL to the image of the article",
                        "type": "string",
                        "example": "https://example.com/article/image.jpg"
                    },
                    "published_at": {
                        "description": "The date and time when the article was published",
                        "type": "string",
                        "example": "2022-05-10 12:00:00"
                    },
                    "created_at": {
                        "description": "The date and time when the user preference was created",
                        "type": "string",
                        "example": "2022-05-10 12:00:00"
                    },
                    "updated_at": {
                        "description": "The date and time when the user preference was last updated",
                        "type": "string",
                        "example": "2022-05-10 12:00:00"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}